+++ eda-modbus-bridge.mjs
@@ -1,5 +1,5 @@
 import express from 'express'
-import morgan from 'morgan'
+import expressWinston from 'express-winston'
 import MQTT from 'async-mqtt'
 import yargs from 'yargs'
 import ModbusRTU from 'modbus-serial'
@@ -12,6 +12,7 @@ import {
     validateBrokerUrl,
 } from './app/mqtt.mjs'
 import { configureMqttDiscovery } from './app/homeassistant.mjs'
+import { createLogger, setLogLevel } from './app/logger.mjs'
 
 const MQTT_INITIAL_RECONNECT_RETRY_INTERVAL_SECONDS = 5
 
@@ -44,7 +45,7 @@ const argv = yargs(process.argv.slice(2))
             alias: 'p',
         },
         'mqttBrokerUrl': {
-            description: 'The URL to the MQTT broker, e.g. tcp://localhost:1883. Omit to disable MQTT support.',
+            description: 'The URL to the MQTT broker, e.g. mqtt://localhost:1883. Omit to disable MQTT support.',
             default: undefined,
             alias: 'm',
         },
@@ -68,11 +69,25 @@ const argv = yargs(process.argv.slice(2))
             type: 'boolean',
             default: true,
         },
+        'debug': {
+            description: 'Enable debug logging',
+            type: 'boolean',
+            default: false,
+            alias: 'v',
+        },
     }).argv
 
 ;(async () => {
+    // Create logger(s)
+    const logger = createLogger('main')
+    if (argv.debug) {
+        setLogLevel(logger, 'debug')
+    }
+
+    const httpLogger = createLogger('http')
+
     // Create Modbus client
-    console.log(`Opening serial connection to ${argv.device}, slave ID ${argv.modbusSlave}`)
+    logger.info(`Opening serial connection to ${argv.device}, slave ID ${argv.modbusSlave}`)
     const modbusClient = new ModbusRTU()
     modbusClient.setID(argv.modbusSlave)
     modbusClient.setTimeout(5000) // 5 seconds
@@ -85,10 +100,12 @@ const argv = yargs(process.argv.slice(2))
 
     // Optionally create HTTP server
     if (argv.http) {
+        // Define middleware
         const httpServer = express()
-        httpServer.use(morgan('tiny'))
+        httpServer.use(expressWinston.logger({ winstonInstance: httpLogger }))
         httpServer.use(express.json())
 
+        // Define routes
         httpServer.get('/', root)
         httpServer.get('/summary', (req, res) => {
             return summary(modbusClient, req, res)
@@ -104,23 +121,23 @@ const argv = yargs(process.argv.slice(2))
         })
 
         httpServer.listen(argv.httpPort, argv.httpListenAddress, () => {
-            console.log(`Listening on http://${argv.httpListenAddress}:${argv.httpPort}`)
+            httpLogger.info(`Listening on http://${argv.httpListenAddress}:${argv.httpPort}`)
         })
     }
 
     // Optionally create MQTT client
     if (argv.mqttBrokerUrl !== undefined) {
         if (!validateBrokerUrl(argv.mqttBrokerUrl)) {
-            console.error(`Malformed MQTT broker URL: ${argv.mqttBrokerUrl}. Should be e.g. tcp://localhost:1883.`)
+            logger.error(`Malformed MQTT broker URL: ${argv.mqttBrokerUrl}. Should be e.g. mqtt://localhost:1883.`)
         } else {
-            console.log(`Connecting to MQTT broker at ${argv.mqttBrokerUrl}`)
+            logger.info(`Connecting to MQTT broker at ${argv.mqttBrokerUrl}`)
 
             try {
                 // Handle authentication
                 let clientOptions = {}
 
                 if (argv.mqttUsername && argv.mqttPassword) {
-                    console.log('Using MQTT broker authentication')
+                    logger.info('Using MQTT broker authentication')
 
                     clientOptions = {
                         'username': argv.mqttUsername,
@@ -138,9 +155,9 @@ const argv = yargs(process.argv.slice(2))
                     try {
                         mqttClient = await MQTT.connectAsync(argv.mqttBrokerUrl, clientOptions)
                         connectedOnce = true
-                        console.log(`Successfully connected to MQTT broker at ${argv.mqttBrokerUrl}`)
+                        logger.info(`Successfully connected to MQTT broker at ${argv.mqttBrokerUrl}`)
                     } catch (e) {
-                        console.error(
+                        logger.error(
                             `Failed to connect to MQTT broker: ${e.message}. Retrying in ${retryIntervalMs} milliseconds`
                         )
 
@@ -156,7 +173,7 @@ const argv = yargs(process.argv.slice(2))
                     await publishValues(modbusClient, mqttClient)
                 }, argv.mqttPublishInterval * 1000)
 
-                console.log(`MQTT scheduler started, will publish readings every ${argv.mqttPublishInterval} seconds`)
+                logger.info(`MQTT scheduler started, will publish readings every ${argv.mqttPublishInterval} seconds`)
 
                 // Subscribe to changes and register a handler
                 await subscribeToChanges(modbusClient, mqttClient)
@@ -167,15 +184,15 @@ const argv = yargs(process.argv.slice(2))
                 // Optionally configure Home Assistant MQTT discovery
                 if (argv.mqttDiscovery) {
                     await configureMqttDiscovery(modbusClient, mqttClient)
-                    console.log('Finished configuration Home Assistant MQTT discovery')
+                    logger.info('Finished configuration Home Assistant MQTT discovery')
                 }
 
                 // Log reconnection attempts
                 mqttClient.on('reconnect', () => {
-                    console.log(`Attempting to reconnect to ${argv.mqttBrokerUrl}`)
+                    logger.info(`Attempting to reconnect to ${argv.mqttBrokerUrl}`)
                 })
             } catch (e) {
-                console.error(`Unknown exception occurred: ${e.message}`, e.stack)
+                logger.error(`An exception occurred: ${e.name}: ${e.message}`, e.stack)
             }
         }
     }