+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## future release
 
 * Drastically improve application logging
+* Add missing error handling to the `/summary` route, change HTTP error responses to return JSON
 
 ## 2.3.2